스터디관리 수정된 레이아웃
